app/requirements.txt
bcrypt==4.2.1
email_validator==2.2.0
fastapi==0.95.2
passlib==1.7.4
pydantic==1.10.8
pydantic_core==2.27.1
PyJWT==2.10.1
SQLAlchemy==1.4.49
uvicorn==0.22.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1